STAR-36105: Restructure API code samples (#14)
This splits the Payment Services examples and the Public API examples into two separate sub-folders to avoid confusion.

I've tidied up the higher level READMEs as necessary and added a couple of missing .gitignore files. All other changes are structural. There are no changes to the examples themselves, nor to the lower level READMEs that relate to the code.

* Starting reorg - key rotation is common to Public APIs and Payment Services

* More reorg - Webhook examples are specific to public APIs

* Moving the Payment services message signing examples

* Moving the public API message signing examples

* Updating main readme

* Updating READMEs at the common, public API and PS API levels.

* More README updates

* Updating READMEs

* More README updates