Contribute a Python example for Webhook verification. (#18)
The existing Java example makes it clear that the signature protocol
is SHA512withRSA but I was not exactly sure what the mechanics of
implementing that should be with Python (e.g. working out that RSA.verify
wants the SHA512 hash of the payload as its first argument) so once I
got it working I thought an example would be nice.